# Frontend/workload-management/src/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8088/api/v1",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/course": {
      "get": {
        "tags": [
          "Course"
        ],
        "operationId": "findAllCourses",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CourseResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Course"
        ],
        "operationId": "saveCourse",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/course/upload": {
      "post": {
        "tags": [
          "Course"
        ],
        "operationId": "uploadCourse",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/auth/forgot-password": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "forgotPassword",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auth/authenticate": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "authenticate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthenticationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/academic-rank": {
      "get": {
        "tags": [
          "AcademicRank"
        ],
        "operationId": "findAllAcademicRank",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AcademicRankResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "AcademicRank"
        ],
        "operationId": "saveAcademicRank",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcademicRankRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/academic-rank/details": {
      "get": {
        "tags": [
          "AcademicRankDetails"
        ],
        "operationId": "findAllAcademicRankDetails",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AcademicRankDetailsResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "AcademicRankDetails"
        ],
        "operationId": "saveAcademicRankDetails",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcademicRankDetailsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
...
    "/course/{courseId}": {
      "get": {
        "tags": [
          "Course"
        ],
        "operationId": "findCourseById",
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Course"
        ],
        "operationId": "deleteCourseById",
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Course"
        ],
        "operationId": "updateCourseById",
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/academic-rank/{academicRankId}": {
      "get": {
        "tags": [
          "AcademicRank"
        ],
        "operationId": "findAcademicRankById",
        "parameters": [
          {
            "name": "academicRankId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademicRankResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "AcademicRank"
        ],
        "operationId": "deleteAcademicRankById",
        "parameters": [
          {
            "name": "academicRankId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "AcademicRank"
        ],
        "operationId": "updateAcademicRankById",
        "parameters": [
          {
            "name": "academicRankId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcademicRankRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/academic-rank/details/{academicRankDetailsId}": {
      "get": {
        "tags": [
          "AcademicRankDetails"
        ],
        "operationId": "findAcademicRankDetailsById",
        "parameters": [
          {
            "name": "academicRankDetailsId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademicRankDetailsResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "AcademicRankDetails"
        ],
        "operationId": "deleteAcademicRankDetailsById",
        "parameters": [
          {
            "name": "academicRankDetailsId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "AcademicRankDetails"
        ],
        "operationId": "updateAcademicRankDetailsById",
        "parameters": [
          {
            "name": "academicRankDetailsId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcademicRankDetailsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/activate-account": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "operationId": "confirm",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auth/activate-account/change-password": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "operationId": "changePassword",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CourseRequest": {
        "type": "object",
        "properties": {
          "courseCode": {
            "type": "string"
          },
          "courseName": {
            "type": "string"
          },
          "creditPoints": {
            "type": "number",
            "format": "double"
          },
          "registrationType": {
            "type": "string"
          },
          "section": {
            "type": "string"
          }
        },
        "required": [
          "courseCode",
          "courseName",
          "creditPoints",
          "registrationType",
          "section"
        ]
      },
      "AuthenticationRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "maxLength": 2147483647,
            "minLength": 8
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "AuthenticationResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
      "AcademicRankRequest": {
        "type": "object",
        "properties": {
          "rankName": {
            "type": "string"
          },
          "abbreviation": {
            "type": "string"
          }
        },
        "required": [
          "abbreviation",
          "rankName"
        ]
      },
      "AcademicRankDetailsRequest": {
        "type": "object",
        "properties": {
          "cpForFullTime": {
            "type": "number",
            "format": "double"
          },
          "salary": {
            "type": "number",
            "format": "double"
          },
          "contactHoursForFullTime": {
            "type": "number",
            "format": "double"
          },
          "semesterId": {
            "type": "integer",
            "format": "int32"
          },
          "academicRankId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "academicRankId",
          "contactHoursForFullTime",
          "cpForFullTime",
          "salary",
          "semesterId"
        ]
      },
      "AcademicRank": {
        "type": "object",
        "properties": {
          "academicRankId": {
            "type": "integer",
            "format": "int32"
          },
          "rankName": {
            "type": "string"
          },
          "abbreviation": {
            "type": "string"
          },
          "deleted": {
            "type": "boolean"
          }
        }
      },
...
      "Semester": {
        "type": "object",
        "properties": {
          "semesterId": {
            "type": "integer",
            "format": "int32"
          },
          "semesterName": {
            "type": "string",
            "enum": [
              "PAVASARIS",
              "RUDENS"
            ]
          },
          "semesterYear": {
            "type": "integer",
            "format": "int32"
          },
          "deleted": {
            "type": "boolean"
          }
        }
      },
...
      "CourseResponse": {
        "type": "object",
        "properties": {
          "courseId": {
            "type": "integer",
            "format": "int32"
          },
          "courseCode": {
            "type": "string"
          },
          "courseName": {
            "type": "string"
          },
          "creditPoints": {
            "type": "number",
            "format": "double"
          },
          "registrationType": {
            "type": "string"
          },
          "section": {
            "type": "string"
          },
          "deleted": {
            "type": "boolean"
          }
        }
      },
      "AcademicRankResponse": {
        "type": "object",
        "properties": {
          "academicRankId": {
            "type": "integer",
            "format": "int32"
          },
          "rankName": {
            "type": "string"
          },
          "abbreviation": {
            "type": "string"
          },
          "deleted": {
            "type": "boolean"
          }
        }
      },
      "AcademicRankDetailsResponse": {
        "type": "object",
        "properties": {
          "academicRankDetailsId": {
            "type": "integer",
            "format": "int32"
          },
          "cpForFullTime": {
            "type": "number",
            "format": "double"
          },
          "salary": {
            "type": "number",
            "format": "double"
          },
          "contactHoursForFullTime": {
            "type": "number",
            "format": "double"
          },
          "semester": {
            "$ref": "#/components/schemas/Semester"
          },
          "academicRank": {
            "$ref": "#/components/schemas/AcademicRank"
          },
          "deleted": {
            "type": "boolean"
          }
        }
      }
    }
  }
}